Reasonable implementation of HTTP/3.
This is the first iteration of Vert.x HTTP/3 implementation which implements partically RFC9114 and discovery via alternative services RFC7838

HTTP push is not implemented as we plan to drop this feature in Vert.x 6.

+++ vertx-core/src/main/java/io/vertx/core/http/HttpClientRequest.java
@@ -535,8 +535,8 @@ public interface HttpClientRequest extends WriteStream<Buffer> {
    * @return the id of the stream of this response, {@literal -1} when it is not yet determined, i.e
    *         the request has not been yet sent or it is not supported HTTP/1.x
    */
-  default int streamId() {
-    return -1;
+  default long streamId() {
+    return -1L;
   }
 
   /**
