+++ vertx-core/src/main/java/io/vertx/core/http/impl/HttpServerResponseImpl.java
@@ -703,7 +703,7 @@ public class HttpServerResponseImpl implements HttpServerResponse {
   }
 
   @Override
-  public int streamId() {
+  public long streamId() {
     return stream.id();
   }
