+++ vertx-core/src/main/java/io/vertx/core/http/impl/http2/DefaultHttp2Stream.java
@@ -252,7 +252,7 @@ abstract class DefaultHttp2Stream<S extends DefaultHttp2Stream<S>> implements Ht
     inboundQueue.write(trailers);
   }
 
-  public final int id() {
+  public final long id() {
     return id;
   }
